# Add performance measurement