@tsmaeder 2021-09-07
@sgraband thanks for this contribution: it's important that we keep track of startup performance. I have a bunch of suggestions/comments which are mostly code style issues.