@sgraband 2021-09-23
@tsmaeder I rebased the changes to master. Could you have another look?