@sgraband 2021-10-06
Rebased and removed the .npmrc file.